api_yamdb/requirements.txt
asgiref==3.5.2
attrs==21.4.0
certifi==2022.6.15
cfgv==3.3.1
charset-normalizer==2.0.12
distlib==0.3.5
Django==2.2.16
django-filter==21.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.2.0
filelock==3.7.1
gunicorn==20.0.4
identify==2.5.2
idna==3.3
importlib-metadata==4.12.0
iniconfig==1.1.1
nodeenv==1.7.0
packaging==21.3
platformdirs==2.5.2
pluggy==0.13.1
pre-commit==2.20.0
psycopg2-binary==2.8.6
py==1.11.0
PyJWT==2.1.0
pyparsing==3.0.9
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python-dotenv==0.21.0
pytz==2022.1
PyYAML==6.0
requests==2.26.0
sqlparse==0.4.2
toml==0.10.2
typing_extensions==4.3.0
urllib3==1.26.11
zipp==3.8.1
